+++ src/Spy/Factory/SpyVerifierFactory.php
@@ -170,7 +170,7 @@ class SpyVerifierFactory implements SpyVerifierFactoryInterface
     /**
      * Create a new spy verifier.
      *
-     * @param SpyInterface|null $spy The spy, or null to create an unbound spy verifier.
+     * @param SpyInterface|null $spy The spy, or null to create an anonymous spy.
      *
      * @return SpyVerifierInterface The newly created spy verifier.
      */
@@ -194,20 +194,20 @@ class SpyVerifierFactory implements SpyVerifierFactoryInterface
     /**
      * Create a new spy verifier for the supplied callback.
      *
-     * @param callable|null $callback            The callback, or null to create an unbound spy verifier.
-     * @param boolean       $useGeneratorSpies   True if generator spies should be used.
-     * @param boolean       $useTraversableSpies True if traversable spies should be used.
+     * @param callable|null $callback The callback, or null to create an anonymous spy.
      *
      * @return SpyVerifierInterface The newly created spy verifier.
      */
-    public function createFromCallback(
-        $callback = null,
-        $useGeneratorSpies = true,
-        $useTraversableSpies = false
-    ) {
-        return $this->create(
-            $this->spyFactory
-                ->create($callback, $useGeneratorSpies, $useTraversableSpies)
+    public function createFromCallback($callback = null)
+    {
+        return new SpyVerifier(
+            $this->spyFactory->create($callback),
+            $this->matcherFactory,
+            $this->matcherVerifier,
+            $this->callVerifierFactory,
+            $this->assertionRecorder,
+            $this->assertionRenderer,
+            $this->invocableInspector
         );
     }