+++ src/Spy/Factory/SpyVerifierFactoryInterface.php
@@ -22,7 +22,7 @@ interface SpyVerifierFactoryInterface
     /**
      * Create a new spy verifier.
      *
-     * @param SpyInterface|null $spy The spy, or null to create an unbound spy verifier.
+     * @param SpyInterface|null $spy The spy, or null to create an anonymous spy.
      *
      * @return SpyVerifierInterface The newly created spy verifier.
      */
@@ -31,15 +31,9 @@ interface SpyVerifierFactoryInterface
     /**
      * Create a new spy verifier for the supplied callback.
      *
-     * @param callable|null $callback            The callback, or null to create an unbound spy verifier.
-     * @param boolean       $useGeneratorSpies   True if generator spies should be used.
-     * @param boolean       $useTraversableSpies True if traversable spies should be used.
+     * @param callable|null $callback The callback, or null to create an anonymous spy.
      *
      * @return SpyVerifierInterface The newly created spy verifier.
      */
-    public function createFromCallback(
-        $callback = null,
-        $useGeneratorSpies = true,
-        $useTraversableSpies = false
-    );
+    public function createFromCallback($callback = null);
 }