+++ src/Stub/Factory/StubFactory.php
@@ -134,7 +134,7 @@ class StubFactory implements StubFactoryInterface
     /**
      * Create a new stub.
      *
-     * @param callable|null $callback              The callback, or null to create an unbound stub.
+     * @param callable|null $callback              The callback, or null to create an anonymous stub.
      * @param mixed         $self                  The self value.
      * @param callable|null $defaultAnswerCallback The callback to use when creating a default answer.
      *