+++ src/Stub/Factory/StubFactoryInterface.php
@@ -21,7 +21,7 @@ interface StubFactoryInterface
     /**
      * Create a new stub.
      *
-     * @param callable|null $callback              The callback, or null to create an unbound stub.
+     * @param callable|null $callback              The callback, or null to create an anonymous stub.
      * @param mixed         $self                  The self value.
      * @param callable|null $defaultAnswerCallback The callback to use when creating a default answer.
      *